models/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile models/requirements.in
#
appdirs==1.4.4
    # via pooch
audioread==3.0.0
    # via librosa
certifi==2022.6.15
    # via requests
cffi==1.15.1
    # via soundfile
charset-normalizer==2.1.1
    # via requests
coloredlogs==15.0.1
    # via onnxruntime
commonmark==0.9.1
    # via rich
cycler==0.11.0
    # via matplotlib
decorator==5.1.1
    # via librosa
distro==1.7.0
    # via scikit-build
facenet-pytorch==2.5.2
    # via -r models/requirements.in
filelock==3.8.0
    # via
    #   huggingface-hub
    #   transformers
flatbuffers==2.0
    # via onnxruntime
fonttools==4.36.0
    # via matplotlib
huggingface-hub==0.9.0
    # via transformers
humanfriendly==10.0
    # via coloredlogs
idna==3.3
    # via requests
joblib==1.1.0
    # via
    #   librosa
    #   scikit-learn
kiwisolver==1.4.4
    # via matplotlib
librosa==0.9.2
    # via -r models/requirements.in
llvmlite==0.39.0
    # via numba
matplotlib==3.5.3
    # via seaborn
mpmath==1.2.1
    # via sympy
numba==0.56.0
    # via
    #   librosa
    #   resampy
numpy==1.22.4
    # via
    #   facenet-pytorch
    #   librosa
    #   matplotlib
    #   numba
    #   onnx
    #   onnxruntime
    #   opencv-python
    #   pandas
    #   resampy
    #   scikit-learn
    #   scipy
    #   seaborn
    #   torchvision
    #   transformers
onnx==1.12.0
    # via
    #   -r models/requirements.in
    #   onnx-simplifier
    #   onnxsim
onnx-simplifier==0.4.7
    # via -r models/requirements.in
onnxruntime==1.12.1
    # via -r models/requirements.in
onnxsim==0.4.7
    # via -r models/requirements.in
opencv-python==4.6.0.66
    # via -r models/requirements.in
packaging==21.3
    # via
    #   huggingface-hub
    #   librosa
    #   matplotlib
    #   onnxruntime
    #   pooch
    #   scikit-build
    #   transformers
pandas==1.4.3
    # via
    #   -r models/requirements.in
    #   seaborn
pillow==9.2.0
    # via
    #   -r models/requirements.in
    #   facenet-pytorch
    #   matplotlib
    #   torchvision
pooch==1.6.0
    # via librosa
protobuf==3.20.1
    # via
    #   onnx
    #   onnxruntime
pycparser==2.21
    # via cffi
pygments==2.13.0
    # via rich
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2022.2.1
    # via pandas
pyyaml==6.0
    # via
    #   huggingface-hub
    #   transformers
regex==2022.8.17
    # via transformers
requests==2.28.1
    # via
    #   facenet-pytorch
    #   huggingface-hub
    #   pooch
    #   torchvision
    #   transformers
resampy==0.4.0
    # via librosa
rich==12.5.1
    # via
    #   onnx-simplifier
    #   onnxsim
scikit-build==0.15.0
    # via -r models/requirements.in
scikit-learn==1.1.2
    # via librosa
scipy==1.9.0
    # via
    #   librosa
    #   scikit-learn
    #   seaborn
seaborn==0.11.2
    # via -r models/requirements.in
six==1.16.0
    # via python-dateutil
soundfile==0.10.3.post1
    # via librosa
sympy==1.10.1
    # via onnxruntime
threadpoolctl==3.1.0
    # via scikit-learn
tokenizers==0.12.1
    # via transformers
torch==1.12.1
    # via
    #   -r models/requirements.in
    #   torchvision
torchvision==0.13.1
    # via
    #   -r models/requirements.in
    #   facenet-pytorch
tqdm==4.64.0
    # via
    #   huggingface-hub
    #   transformers
transformers==4.21.1
    # via -r models/requirements.in
typing-extensions==4.3.0
    # via
    #   huggingface-hub
    #   onnx
    #   torch
    #   torchvision
urllib3==1.26.12
    # via requests
wheel==0.37.1
    # via scikit-build

# The following packages are considered to be unsafe in a requirements file:
# setuptools
